refactor: rename files and classes from ...Filters to ...Filter
