+++ apps/web/src/utils/program/programFilter.ts
